Add files via upload
Makes a count of the available electric vehicles for each hour. Travel patterns were taken from the National Household Travel Survey. For 2023 we estimate that there are 67,444 BEVs on the road in San Diego (California Energy Commission). For 2030 we project there to be 337,050 BEVs on the road (San Diego County EV Roadmap)